Upgrade expo sdk version to 53

+++ src/utils/StorageManager/migrations/addReuseCountMigration.ts
@@ -2,16 +2,24 @@ import { MigrationInterface, QueryRunner } from 'typeorm';
 
 export class AddReuseCountColumn163837490194410 implements MigrationInterface {
     public async up(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.query(`
+        const tableExists = await queryRunner.hasTable('IdentityVerificationStorage');
+
+        if (tableExists) {
+            await queryRunner.query(`
             ALTER TABLE "IdentityVerificationStorage" 
             ADD COLUMN "reuseCount" integer NOT NULL DEFAULT 0
-        `);
+            `);
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.query(`
-            ALTER TABLE "IdentityVerificationStorage" 
-            DROP COLUMN "reuseCount"
-        `);
+        const tableExists = await queryRunner.hasTable('IdentityVerificationStorage');
+
+        if (tableExists) {
+            await queryRunner.query(`
+                ALTER TABLE "IdentityVerificationStorage" 
+                DROP COLUMN "reuseCount"
+            `);
+        }
     }
 }
